// apps/web/src/api/generated/orval/model/chatMessageResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ChatMessageResponseMessageType } from './chatMessageResponseMessageType';
import type { ChatMessageResponseRole } from './chatMessageResponseRole';
import type { ChatMessageResponseStatus } from './chatMessageResponseStatus';
import type { JsonNode } from './jsonNode';

/**
 * 챗 메시지
 */
export interface ChatMessageResponse {
  /** 메시지 ID */
  message_id?: string;
  /** 실행 ID */
  run_id?: string;
  /** 메시지 역할 */
  role?: ChatMessageResponseRole;
  /** 메시지 타입 */
  message_type?: ChatMessageResponseMessageType;
  /** 메시지 상태 */
  status?: ChatMessageResponseStatus;
  /** 메시지 순서 */
  sequence?: number;
  /** 메시지 내용 */
  content?: JsonNode;
  /** 생성 시각 */
  created_at?: string;
}